cloudcam-bridge/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiofiles==23.2.1
pyyaml==6.0.1
requests==2.31.0
